requirements.txt
streamlit
pdfplumber
pytesseract
Pillow
pandas
openpyxl
requests
